pip-requirements.txt
requests>=1.1.0
ckantoolkit
